Rejecting and flow control on PromiseIterator
The resolve() method of PromiseIterator returns a Promise!bool which
will be fulfilled once the each callback is done with the given value.
Now, TcpSocket uses this promise to do flow control on the input data.
It will read the next TCP buffer only when the previous one is ready.
Also, the PromiseIterator might reject right away, which will propagate
the exception after the each method.