Better handling of void Promise type
Before this commit, Promise!void was just a alias for Promise!void*.
Now, PromiseBase!T... was created. Having variadic arguments make it
possible to have zero arguments, which is the cause of the void.
The classes Promise!T and Promise!void are subclasses of PromiseBase
with methods to simplify the meta-programming argument matching of the compiler on the
resolve() and then() function. Without this simplification, the user
would have to always explicify the types involved in a callback.