Created finally concept for Promise
The finally block is a block that is always called: It is called when
an exception is thrown and by the normal flow. Finally exception can
itself throw another exception or return a new value. If it does not
throw an exception and an exception was already thrown, the thrown
exception will continue the flow till a except block is found. The
exception that the finally block throws will override the current
exception.
Also, did some refactors that improved code stability as alternative
if-then-else were minimized.