Created ReadoneStream concept
ReadoneStream is a abstract implementation of Stream which is based in
reading one buffer per time. Subclasses of ReadoneStream must implement
readOne() method that will then call readOneData or readOneReject
indicating the result of a single read. The ReadoneStream will be
responsible for calling readOne N times regarding the calls to read()
and if the iteration as interrupted by the result of the each()
callback.