Interface Loop avoids external include dependency
Creates the interface Loop that declares functionality to resolve a hostname, to connect to a TCP host, and to do client side TLS handshaking.
The default implementation will rely on libuv and OpenSSL.
The code that includes the upromised.loop module does not need to have OpenSSL or LibUV in the include path.