Promise and PromiseIterator are interfaces
Remade the Promise core implementation to work based as interfaces.
The Promise interface has only one abstract method that is `then_` that registers a callback to the Promise. The PromiseIterator has only one abstract method that is `next` that returns a Promise of the next element in the iteration.
Buffering the resolve is made by the subclass DelegatePromise, and buffering the iteration is done by the DelegatePromiseIterator subclass.
The new implementation is more simple and easier to validate.