Added proxy command example
proxy command receive two args list (splited by --). The first is the proxy command, the command that will do the actual connection (using netcat protocol). The second command is the command that will be invoked, %h and %p will be replaced for the hostname and port of a TCP server that will proxy the connection using the specified proxy command.