Not using GC.malloc for buffer allocation on ArmHF
On LDC2 for Arm-HardFloat, allocating memory for a buffer using GC.malloc is leaking memory. For some reason, the buffers are never collected. Switching that to a new array expression fixes the problem. Since realloc for an array is not available, the shrinkBuf function was modified to create a copy of the buffer in a smaller array.